+++ R/sendrecv.R
@@ -44,12 +44,15 @@
 #'     queued for sending. Certain protocol / transport combinations may limit
 #'     the number of messages that can be queued if they have yet to be received.
 #'
-#'     For Contexts and Streams: the default behaviour is blocking with
-#'     \code{block = TRUE}. This will wait until the send has completed. Set a
-#'     timeout in this case to ensure that the function returns under all scenarios.
-#'     As the underlying implementation uses an asynchronous send with a wait,
-#'     it is recommended to set a positive integer value for \code{block} rather
-#'     than FALSE.
+#'     For Contexts: the default behaviour is blocking with \code{block = TRUE}.
+#'     This will wait until the send has completed. Set to FALSE or an integer
+#'     timeout to ensure that the function returns under all scenarios.
+#'
+#'     For Streams: the default behaviour is blocking with \code{block = TRUE}.
+#'     This will wait until the send has completed. Set a timeout to ensure that
+#'     the function returns under all scenarios. As the underlying implementation
+#'     uses an asynchronous send with a wait, it is recommended to set a positive
+#'     integer value for \code{block} rather than FALSE.
 #'
 #' @seealso \code{\link{send_aio}} for asynchronous send.
 #' @examples