+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.10.0.9005 (development)
+# nanonext 0.10.0.9006 (development)
 
 #### New Features